+++ src/server/newscripts/Pandaria/SiegeofOrgrimmar/boss_spoils_of_pandaria.cpp
@@ -1814,7 +1814,7 @@ class spell_rage_of_the_empress : public SpellScriptLoader
         {
             PrepareSpellScript(spell_rage_of_the_empress_SpellScript);
 
-            int32 bp0;
+            float bp0;
             bool Load()
             {
                 bp0 = 0;
@@ -2084,7 +2084,7 @@ class spell_boss_torment : public SpellScriptLoader
                 if (!caster)
                     return;
 
-                int32 amount = aurEff->GetAmount() + int32(aurEff->GetAmount() * 0.1f * aurEff->GetTickNumber());
+                float amount = aurEff->GetAmount() + int32(aurEff->GetAmount() * 0.1f * aurEff->GetTickNumber());
                 caster->CastCustomSpell(target, 136885, &amount, 0, 0, true, 0, aurEff);
             }
